+++ lib/src/enum/message_type.dart
@@ -48,6 +48,10 @@ class MessageType {
   /// 高级撤回
   static const advancedRevoke = 118;
 
+  static const customMsgNotTriggerConversation = 119;
+
+  static const customMsgOnlineOnly = 120;
+
   /// 通知类型
   static const notificationBegin = 1000;